requirements.txt
anthropic>=0.3.0
anthropic>=0.3.0
flask
flask-cors
anthropic
python-dotenv
gunicorn